Rename package and component
