+++ demo/src/pages/markdown.mdx
@@ -3,7 +3,7 @@ title: Components in MDX
 layout: ../layouts/Base.astro
 ---
 
-import { Tweet, YouTube, OpenGraph } from 'astro-embed';
+import { Tweet, YouTube, LinkPreview } from 'astro-embed';
 
 The embeds on this page are imported as components in the MDX frontmatter’s `setup` block and then used in the document body.
 
@@ -24,6 +24,6 @@ setup: |
 
     <YouTube id="Hoe-woAhq_k" />
 
-## `<OpenGraph />`
+## `<LinkPreview />`
 
-    <OpenGraph id="https://astro.build/blog/welcome-world/" />
+    <LinkPreview id="https://astro.build/blog/welcome-world/" />
