+++ packages/astro-embed-integration/package.json
@@ -22,7 +22,7 @@
   },
   "homepage": "https://astro-embed.netlify.app/getting-started/#auto-embed-urls-in-mdx",
   "dependencies": {
-    "@astro-community/astro-embed-og": "^0.0.1",
+    "@astro-community/astro-embed-link-preview": "^0.0.1",
     "@astro-community/astro-embed-twitter": "^0.5.4",
     "@astro-community/astro-embed-vimeo": "^0.3.4",
     "@astro-community/astro-embed-youtube": "^0.5.0",
