+++ packages/astro-embed-link-preview/index.ts
@@ -1,0 +1,1 @@
+export { default as LinkPreview } from './LinkPreview.astro';
