+++ packages/astro-embed-link-preview/package.json
@@ -1,28 +1,33 @@
 {
-  "name": "@astro-community/astro-embed-og",
+  "name": "@astro-community/astro-embed-link-preview",
   "version": "0.0.1",
-  "description": "Component to easily embed websites via social metatags on your Astro site",
+  "description": "Component to embed a website’s OpenGraph image and metadata on your Astro site",
   "type": "module",
   "exports": {
     ".": "./index.ts",
-    "./lib": "./lib.ts",
     "./matcher": "./matcher.ts"
   },
   "files": [
     "lib.ts",
     "index.ts",
     "matcher.ts",
-    "OpenGraph.astro"
+    "LinkPreview.astro"
   ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/delucis/astro-embed.git"
   },
+  "keywords": [
+    "astro",
+    "astro-component",
+    "embeds",
+    "open-graph"
+  ],
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/delucis/astro-embed/issues"
   },
-  "homepage": "https://github.com/delucis/astro-embed/tree/main/packages/astro-embed-og#readme",
+  "homepage": "https://github.com/delucis/astro-embed/tree/main/packages/astro-embed-link-preview#readme",
   "dependencies": {
     "@astro-community/astro-embed-utils": "^0.1.0"
   }
