--- packages/astro-embed-og/index.ts
@@ -1,1 +1,0 @@
-export { default as OpenGraph } from './OpenGraph.astro';
