+++ packages/astro-embed/index.ts
@@ -1,4 +1,4 @@
 export { Tweet } from '@astro-community/astro-embed-twitter';
 export { YouTube } from '@astro-community/astro-embed-youtube';
 export { Vimeo } from '@astro-community/astro-embed-vimeo';
-export { OpenGraph } from '@astro-community/astro-embed-og';
+export { LinkPreview } from '@astro-community/astro-embed-link-preview';
